move processWebhookMessage to async
