+++ apps/meteor/app/lib/server/functions/index.ts
@@ -15,7 +15,6 @@ export { generateUsernameSuggestion } from './getUsernameSuggestion';
 export { insertMessage } from './insertMessage';
 export { isTheLastMessage } from './isTheLastMessage';
 export { loadMessageHistory } from './loadMessageHistory';
-export { processWebhookMessage } from './processWebhookMessage';
 export { removeUserFromRoom } from './removeUserFromRoom';
 export { relinquishRoomOwnerships } from './relinquishRoomOwnerships';
 export { saveCustomFields } from './saveCustomFields';
