+++ apps/meteor/app/lib/server/functions/processWebhookMessage.js
@@ -7,7 +7,7 @@ import { validateRoomMessagePermissions } from '../../../authorization/server/fu
 import { SystemLogger } from '../../../../server/lib/logger/system';
 import { trim } from '../../../../lib/utils/stringUtils';
 
-export const processWebhookMessage = function (messageObj, user, defaultValues = { channel: '', alias: '', avatar: '', emoji: '' }) {
+export const processWebhookMessage = async function (messageObj, user, defaultValues = { channel: '', alias: '', avatar: '', emoji: '' }) {
 	const sentData = [];
 	const channels = [].concat(messageObj.channel || messageObj.roomId || defaultValues.channel);
 
@@ -101,7 +101,7 @@ export const processWebhookMessage = function (messageObj, user, defaultValues =
 
 		validateRoomMessagePermissions(room, { uid: user._id, ...user });
 
-		const messageReturn = Promise.await(sendMessage(user, message, room));
+		const messageReturn = await sendMessage(user, message, room);
 		sentData.push({ channel, message: messageReturn });
 	}
